Implement macros defined in <stdbool.h>
Macro 'bool', 'true', 'false' are defined in c.c. In addition, this
commit also tweaks 'define' preprocessor directive so that parsing
identifier after 'define' is possible.

The size of type _Bool is same as char type, which takes 1 byte.

Notice that current lexer aliasing is unreliable due to the token
type determination is hardcoded, identifiers other than `_Bool` will
not be distinguished as T_identifier but T_string.