Experiments with inliner changes
This patch initially moves some definitions to their original header
files defined in C89 or later specifications, and also did some changes
on inliner.c and Makefile so that it can takes more than 1 files to be
inlined with into the produced shecc binary.

Now the only problem with this approach is that some times declarations
is defined in multiple places and inliner cannot actually perform macro
expansion, which results in forbidden usage of any C macro features, and
probably would make this approach useless. Right now this patch is
buildable, but for the final goal, it's not gonna work, so I'll just
leave the works here and head to make frontend improvements so we can
later revisit this patch.